data: re-initialize OHLC with 5 years of yfinance data

+++ app/api/routes/ohlc.py
@@ -18,7 +18,7 @@ class OHLCRecord(BaseModel):
     high: float
     low: float
     close: float
-    volume: int
+    volume: float  # Changed from int to float to handle database values
 
 
 class OHLCResponse(BaseModel):
